requirements.txt
kafka-python==2.0.1
# confluent-kafka==1.4.2
python-rapidjson==0.9.1 
geventhttpclient==1.4.2
rtsp==1.1.8 